ajustes finos usb conexão v3

+++ app/src/main/java/com/documentos/wms_beirario/model/recebimentoRfid/ResponseSearchDetailsEpc.kt
@@ -12,7 +12,7 @@ data class ResponseSearchDetailsEpc(
     val nomeProduto: String,
     val notaFiscal: String,
     val numeroSerie: String,
-    val quantidade: Int,
+    val quantidade: Double,
     val tamanho: String,
     val unidadeMedida: String,
     val dataEmissao: String,
